% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRowEach.R
\name{addRowEach}
\alias{addRowEach}
\title{addRowEach}
\usage{
addRowEach(data_frame, interval = 4)
}
\arguments{
\item{data_frame}{Dataframe contenant les donnees du excel}

\item{interval}{Nombre de lignes avant de rajouter une ligne vide}
}
\value{
Renvoi une data.frame avec les lignes vides supplementaires
}
\description{
addRowEach
}
