% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLineBreakToSpace.R
\name{convertLineBreakToSpace}
\alias{convertLineBreakToSpace}
\title{convertLineBreakToSpace}
\usage{
convertLineBreakToSpace(text)
}
\arguments{
\item{text}{Chaine de caracteres}
}
\value{
Renvoi la chaine de caractere sans les saut de lignes
}
\description{
convertLineBreakToSpace
}
