% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linetypeArg.R
\name{linetypeArg}
\alias{linetypeArg}
\title{linetypeArg}
\usage{
linetypeArg(couleur = "blue")
}
\arguments{
\item{couleur}{Choix de la couleur d'affichage}
}
\value{
Renvoi differentes valeurs utilisables dans l'argument linetype de ggplot
}
\description{
linetypeArg
}
