% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeArg.R
\name{shapeArg}
\alias{shapeArg}
\title{shapeArg}
\usage{
shapeArg(couleur = "blue")
}
\arguments{
\item{couleur}{Choix de la couleur d'affichage}
}
\value{
graph montrant les differentes possibilites
}
\description{
shapeArg
}
