% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcopWeightConvert.R
\name{simcopWeightConvert}
\alias{simcopWeightConvert}
\title{simcopWeightConvert}
\usage{
simcopWeightConvert(..., nbClasse = 5)
}
\arguments{
\item{...}{number separate with ","}

\item{nbClasse}{nombre de classe utilisees par eclaircie}
}
\value{
Renvoi les nombres formattes au mode script
}
\description{
simcopWeightConvert
}
