% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyString.R
\name{tidyString}
\alias{tidyString}
\title{tidyString}
\usage{
tidyString(text)
}
\arguments{
\item{text}{Chaine de caractere}
}
\value{
Renvoi une chaine de caractere sans majuscules, caracteres ASCII, caracteres speciaux et remplace les espaces par des "_"
}
\description{
tidyString
}
